Consolidate graph visualizers
